Extending Prometheus metrics by hardware metrics: - machine_cpu_physical_cores - machine_cpu_sockets - machine_dimm_capacity_bytes - machine_dimm_count - machine_nvm_capacity
