+++ metrics/metrics.go
@@ -1,0 +1,42 @@
+// Copyright 2020 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package metrics
+
+import (
+	"time"
+
+	info "github.com/google/cadvisor/info/v1"
+)
+
+// metricValue describes a single metric value for a given set of label values
+// within a parent containerMetric.
+type metricValue struct {
+	value     float64
+	labels    []string
+	timestamp time.Time
+}
+
+type metricValues []metricValue
+
+// infoProvider will usually be manager.Manager, but can be swapped out for testing.
+type infoProvider interface {
+	// SubcontainersInfo provides information about all subcontainers of the
+	// specified container including itself.
+	SubcontainersInfo(containerName string, query *info.ContainerInfoRequest) ([]*info.ContainerInfo, error)
+	// GetVersionInfo provides information about the version.
+	GetVersionInfo() (*info.VersionInfo, error)
+	// GetMachineInfo provides information about the machine.
+	GetMachineInfo() (*info.MachineInfo, error)
+}
